refactor: 폴더 구조 변경 (#96)
- queries 폴더 구조 변경
