+++ src/api/queries/enrolled-student/useGetAvailableCourses.ts
