+++ src/api/queries/enrolled-student/useGetMyEnrolledCourses.ts
@@ -1,5 +1,5 @@
+import { getMyEnrolledCourses } from '@/api/mypage'
 import { useQuery } from '@tanstack/react-query'
-import { getMyEnrolledCourses } from '../mypage'
 
 export const MY_ENROLLED_COURSES_QUERY_KEY = ['my-enrolled-courses']
