+++ src/api/queries/myInfo/useChangePassword.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query'
 import type { ChangePasswordResponse } from '@/types/api-response/myPageResponse'
 import type { ChangePasswordRequest } from '@/types/api-request/myPageRequest'
-import { postChangePassword } from '../mypage'
+import { postChangePassword } from '@/api/mypage'
 
 export const useChangePassword = () => {
   return useMutation<ChangePasswordResponse, Error, ChangePasswordRequest>({
