+++ src/api/queries/myInfo/useChangePhone.ts
@@ -1,5 +1,5 @@
+import { changePhone } from '@/api/mypage'
 import { useMutation } from '@tanstack/react-query'
-import { changePhone } from '../mypage'
 
 export const useChangePhone = () => {
   return useMutation({
