+++ src/features/mypage/MyInfoTab.tsx
@@ -3,9 +3,9 @@ import { useState } from 'react'
 import MyInfoView from './MyInfoView'
 import MyInfoEdit from './MyInfoEdit'
 import { useGetMyInfo } from '@/api/queries/useGetMyInfo'
-import { useGetMyEnrolledCourses } from '@/api/queries/useGetMyEnrolledCourses'
 import { usePatchMyInfo } from '@/api/queries/usePatchMyInfo'
 import Loading from '@/components/common/loading/Loading'
+import { useGetMyEnrolledCourses } from '@/api/queries/enrolled-student/useGetMyEnrolledCourses'
 
 export default function MyInfoTab() {
   const [isEditMode, setIsEditMode] = useState(false)
